finish part5

+++ p1-rasterizer/src/drawrend.h
@@ -63,7 +63,7 @@ class DrawRend : public Renderer {
                            Color color, Triangle *tri = NULL );
 
   //for part4额外添加一个计算重心坐标系数的函数
-  bool bary_coord(int x, int y, float x0, float y0, float x1, float y1, float x2, float y2, float* params);
+  bool bary_coord(float x, float y, float x0, float y0, float x1, float y1, float x2, float y2, float* params);
 
 
 private:
